initial phantom aws installation
Follows the format used by existing modules to now include an
installation of Phantom on a fresh CentOS 7 instance using terraform and
ansible. Adds the needed configuration and a 1 or 0 setting to determine
if the user wants to install Phantom.

Additional plumbing to connect Phantom to Splunk and the Windows
instances will come later.